requirements.txt
fastapi
uvicorn
motor
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
jose
pydantic[email]
httpx
pytest
fastapi-mail
Jinja2